requirements.txt
arxiv>=2.1.0
feedparser>=6.0.11
httpx>=0.27.0
PyYAML>=6.0.1
python-dateutil>=2.9.0
tenacity>=8.2.3
openai>=1.40.0
anthropic>=0.34.0
